add like feature

+++ src/fetaures/posts/actions/deletePost.ts
@@ -3,6 +3,7 @@ import { prisma } from "@/lib/prisma"
 import { POSTS } from "@/path";
 import { revalidatePath } from "next/cache";
 import {cacheSession} from "@/lib/session";
+import {deleteImages} from "@/fetaures/posts/actions/deleteImage";
 
 
 
@@ -11,12 +12,20 @@ export const deletePost = async(id:string):Promise<void> =>{
     if (!session){
         throw new Error("unauthorized to delete")
     }
-
-try {
-    await prisma.posts.delete({
+    try {
+    const post = await prisma.posts.findFirst({
         where:{
             id,
             userId:session.user.id
+        }
+    })
+    if (!post)throw new Error("post not found")
+    if (post.images.length>0) {
+        await deleteImages(post.images)
+    }
+    await prisma.posts.delete({
+        where:{
+            id
         }
     })
     revalidatePath(POSTS);
